remove legacy rich_text, rename rich_text2
