+++ client/apps/edit/components/content/sections/video/index.jsx
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types'
-import Paragraph from '../../../../../../components/rich_text2/components/paragraph.coffee'
+import Paragraph from '../../../../../../components/rich_text/components/paragraph.coffee'
 import React, { Component } from 'react'
 import { Controls } from './controls.jsx'
 import { Video, IconRemove } from '@artsy/reaction-force/dist/Components/Publishing'
