+++ client/apps/settings/client/curations/gucci/test/series.test.js
@@ -4,7 +4,7 @@ import { SeriesAdmin } from '../components/series.jsx'
 import { mount } from 'enzyme'
 
 import ImageUpload from 'client/apps/edit/components/admin/components/image_upload.coffee'
-import Paragraph from 'client/components/rich_text2/components/paragraph.coffee'
+import Paragraph from 'client/components/rich_text/components/paragraph.coffee'
 
 describe('Series Admin', () => {
   let props
