+++ client/assets/main.styl
@@ -1,19 +1,19 @@
 @import '../../node_modules/artsy-ezel-components/pagination'
 @import '../components/layout/stylesheets'
+
+@import '../components/admin_form'
+@import '../components/article_list'
 @import '../components/autocomplete'
 @import '../components/autocomplete_select'
-@import '../components/paginated_list'
-@import '../components/image_upload_form'
-@import '../components/admin_form'
-@import '../components/file_input'
-@import '../components/filter_search'
-@import '../components/article_list'
-@import '../components/rich_text'
-@import '../components/rich_text2'
-@import '../components/rich_text_caption'
-@import '../components/tag_list'
 @import '../components/drag_drop'
 @import '../components/drop_down'
+@import '../components/file_input'
+@import '../components/filter_search'
+@import '../components/image_upload_form'
+@import '../components/paginated_list'
+@import '../components/rich_text'
+@import '../components/tag_list'
+
 @import '../apps/edit'
 @import '../apps/settings/stylesheets'
 @import '../apps/queue'
