+++ client/components/character_limit/index.jsx
@@ -1,5 +1,5 @@
-import Paragraph from 'client/components/rich_text2/components/paragraph.coffee'
-import { PlainText } from '/client/components/rich_text2/components/plain_text.jsx'
+import Paragraph from 'client/components/rich_text/components/paragraph.coffee'
+import { PlainText } from '/client/components/rich_text/components/plain_text.jsx'
 import PropTypes from 'prop-types'
 import React from 'react'
