+++ client/components/rich_text/components/nav.coffee
