+++ client/components/rich_text/components/paragraph.coffee
