+++ client/components/rich_text/test/paragraph.test.coffee
