+++ client/components/rich_text/test/plain_text.test.js
