+++ client/components/rich_text/utils/config.coffee
