+++ client/components/rich_text/utils/decorators.coffee
@@ -35,3 +35,28 @@ exports.Link = (props) ->
   return (
     link
   )
+
+exports.findContentStartEntities = (contentBlock, callback, contentState) ->
+  contentBlock.findEntityRanges ((character) ->
+    entityKey = character.getEntity()
+    entityKey isnt null and contentState.getEntity(entityKey).getType() is 'CONTENT-START'
+  ), callback
+  return
+
+exports.findContentEndEntities = (contentBlock, callback, contentState) ->
+  contentBlock.findEntityRanges ((character) ->
+    entityKey = character.getEntity()
+    entityKey isnt null and contentState.getEntity(entityKey).getType() is 'CONTENT-END'
+  ), callback
+  return
+
+exports.ContentStartEnd = (props) ->
+  type = props.contentState.getEntity(props.entityKey).getType()
+  if type is 'CONTENT-START'
+    return span {
+      className: type.toLowerCase()
+    }, props.children
+  else if type is 'CONTENT-END'
+    return span {
+      className: type.toLowerCase()
+    }
