+++ client/components/rich_text/utils/icons.jade
@@ -1,3 +1,7 @@
+.artist
+  include ../../../apps/edit/public/icons/edit_text_follow_link.svg
+.blockquote
+  include ../../../apps/edit/public/icons/edit_text_blockquote.svg
 .link
   include ../../../apps/edit/public/icons/edit_text_link.svg
 .remove
@@ -8,7 +12,3 @@
   include ../../../apps/edit/public/icons/edit_text_ol.svg
 .unordered-list-item
   include ../../../apps/edit/public/icons/edit_text_ul.svg
-.artist
-  include ../../../apps/edit/public/icons/edit_text_follow_link.svg
-.blockquote
-  include ../../../apps/edit/public/icons/edit_text_blockquote.svg
