--- client/components/rich_text2/index.styl
@@ -1,112 +1,0 @@
-.rich-text--paragraph
-  position relative
-
-.rich-text--url-input
-  position absolute
-  background-color black
-  height 50px
-  width 400px
-  padding 10px
-  display flex
-  input
-    background-color white
-    width 300px
-    max-width 300px
-    garamond l-caption
-    height 30px
-  svg
-    width 35px
-    height 35px
-    position absolute
-    right 90px
-    top 8px
-    cursor pointer
-    z-index 100
-    &:hover polygon
-      fill red-color
-      transition color .15s
-  button.remove-link
-    margin-left -30px
-    width 30px
-    background none transparent
-    border none
-  button.add-link
-    background gray-lighter-color
-    border 0
-    padding 0
-    width 80px
-    height 30px
-    text-align center
-    avant-garde s-headline
-    transition color .15s
-    &:hover
-      color purple-color
-  button.add-link::after
-    content ''
-    width 0
-    height 0
-    border-left 7px solid transparent
-    border-right 7px solid transparent
-    border-bottom 7px solid #000
-    position absolute
-    top -7px
-    left 50%
-
-.rich-text--nav
-  position absolute
-  top -40px
-  left 0
-  display flex !important
-  flex-wrap wrap
-  background black
-  justify-content center
-  border-radius 5px
-  z-index 6
-  ::after
-    content ''
-    width 0
-    height 0
-    border-left 7px solid transparent
-    border-right 7px solid transparent
-    border-top 7px solid #000
-    position absolute
-    bottom -7px
-    left calc(50% \- 7px)
-  button
-    font-weight 600
-    height 40px
-    width 50px
-    color gray-dark-color
-    line-height 1em !important
-    background none
-    border none
-    border-right 1px solid gray-darkest-color
-    transition color .15s
-    &:last-child
-      border-right none
-    &:hover
-      color gray-color
-    &[name='ITALIC'],
-    &[name='STRIKETHROUGH'],
-    &[name='header-two'],
-    &[name='header-three']
-      garamond l-body
-    &[name='header-one']
-      Unica s19
-      font-size 18px
-    &[name='BOLD']
-      font-weight 600
-      avant-garde l-headline
-    &[name='STRIKETHROUGH']
-      text-decoration line-through
-    svg
-      width 18px
-      height 18px
-    .st0, circle, g, path
-      fill gray-dark-color
-      transition color .15s
-    &:hover
-      .st0, circle, g, path
-        fill gray-color
-      &:hover g
-        stroke gray-color
