--- client/components/rich_text2/utils/icons.jade
@@ -1,14 +1,0 @@
-.artist
-  include ../../../apps/edit/public/icons/edit_text_follow_link.svg
-.blockquote
-  include ../../../apps/edit/public/icons/edit_text_blockquote.svg
-.link
-  include ../../../apps/edit/public/icons/edit_text_link.svg
-.remove
-  include ../../../apps/edit/public/icons/edit_text_link_remove.svg
-.remove-formatting
-  include ../../../apps/edit/public/icons/edit_text_remove_formatting.svg
-.ordered-list-item
-  include ../../../apps/edit/public/icons/edit_text_ol.svg
-.unordered-list-item
-  include ../../../apps/edit/public/icons/edit_text_ul.svg
