--- client/components/rich_text_caption/index.styl
@@ -1,47 +1,0 @@
-@import '../stylus_lib'
-
-.rich-text--caption
-  display flex
-  justify-content flex-start
-  position relative
-  &__input
-    garamond l-caption
-    width 100%
-    color gray-darker-color
-  .public-DraftEditorPlaceholder-root
-    position absolute
-    color gray-dark-color
-    left 0
-    right 0
-  .public-DraftEditor-content
-    user-select text
-  &__actions
-    height 40px
-    width 85px
-    background-color black
-    display flex
-    justify-content space-around
-    button
-      width 50%
-      background none transparent
-      outline none
-      border none
-      padding 10px
-      color gray-dark-color
-      &:hover
-        color gray-color
-        svg path
-          fill gray-color
-      svg
-        width 14px
-        height 14px
-        path
-          fill gray-dark-color
-          transition all 0.15s
-    .italic
-      garamond l-body
-      font-style italic
-      font-weight 600
-      line-height 1em
-      border-right 1px solid gray-darkest-color
-      width calc(50% \- 1px)
